Re-balance test list in Jenkins
* run jetstream only for release-32bit
* run test262 by release mode in 64bit
* for test262, add --fast-mode to run the TCs fastly in debug mode
* fix minor bug about type transition

+++ src/runtime/GlobalObjectBuiltinString.cpp
@@ -95,8 +95,9 @@ static Value builtinStringIndexOf(ExecutionState& state, Value thisValue, size_t
         pos = 0;
     }
     size_t len = str->length();
-    size_t start = std::min((size_t)std::max(pos, (double)0), len);
+    size_t start = std::min(std::max(pos, 0.0), (double)len);
     size_t result = str->find(searchStr, start);
+
     if (result == SIZE_MAX)
         return Value(-1);
     else
